+++ projects/jrng-ui/popout/public-api.ts
@@ -1,0 +1,4 @@
+export * from './popout.types';
+export * from './popout-ref';
+export * from './popout.component';
+export * from './popout.service';